requirements.txt
flask==2.0.1
pymongo==4.1.1
nltk==3.6.3
google-api-python-client==2.47.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
networkx==2.6.3
matplotlib==3.4.3
python-dotenv==0.19.0
flask-cors==3.0.10 